Extend the netcdf API to support programmatic changes to the plugin search path

+++ libnczarr/zinternal.c
@@ -113,7 +113,7 @@ find_var_dim_max_length(NC_GRP_INFO_T *grp, int varid, int dimid,
     *maxlen = 0;
 
     /* Find this var. */
-    var = (NC_VAR_INFO_T*)ncindexith(grp->vars,varid);
+    var = (NC_VAR_INFO_T*)ncindexith(grp->vars,(size_t)varid);
     if (!var) return NC_ENOTVAR;
     assert(var->hdr.id == varid);
 
@@ -230,7 +230,7 @@ ncz_find_dim_len(NC_GRP_INFO_T *grp, int dimid, size_t **len)
 {
     NC_VAR_INFO_T *var;
     int retval;
-    int i;
+    size_t i;
 
     assert(grp && len);
     LOG((3, "%s: grp->name %s dimid %d", __func__, grp->hdr.name, dimid));
@@ -382,7 +382,7 @@ static int
 close_dims(NC_GRP_INFO_T *grp)
 {
     NC_DIM_INFO_T *dim;
-    int i;
+    size_t i;
 
     for (i = 0; i < ncindexsize(grp->dim); i++)
     {
@@ -417,7 +417,7 @@ close_dims(NC_GRP_INFO_T *grp)
 static int
 close_types(NC_GRP_INFO_T *grp)
 {
-    int i;
+    size_t i;
 
     for (i = 0; i < ncindexsize(grp->type); i++)
     {
@@ -456,7 +456,7 @@ close_types(NC_GRP_INFO_T *grp)
 static int
 ncz_rec_grp_NCZ_del(NC_GRP_INFO_T *grp)
 {
-    int i;
+    size_t i;
     int retval;
 
     assert(grp && grp->format_grp_info);
@@ -608,7 +608,7 @@ ncz_find_grp_var_att(int ncid, int varid, const char *name, int attnum,
     if (att)
     {
         my_att = use_name ? (NC_ATT_INFO_T *)ncindexlookup(attlist, my_norm_name) :
-            (NC_ATT_INFO_T *)ncindexith(attlist, attnum);
+            (NC_ATT_INFO_T *)ncindexith(attlist, (size_t)attnum);
         if (!my_att)
             return NC_ENOTATT;
     }
